refactor: move singletons
