+++ cognite_toolkit/_cdf_tk/_parameters/type_hint.py
@@ -106,54 +106,3 @@ class TypeHint:
             if self._is_none_type(arg):
                 continue
             yield TypeHint(arg)
-
-
-class _AnyInt(int):
-    def __eq__(self, other: object) -> bool:
-        if isinstance(other, int):
-            return True
-        return NotImplemented
-
-    def __hash__(self) -> int:
-        return id(ANY_INT)
-
-    def __str__(self) -> str:
-        return "AnyInt"
-
-    def __repr__(self) -> str:
-        return "AnyInt"
-
-
-class _AnyStr(str):
-    def __eq__(self, other: object) -> bool:
-        if isinstance(other, str):
-            return True
-        return NotImplemented
-
-    def __hash__(self) -> int:
-        return id(ANY_STR)
-
-    def __str__(self) -> str:
-        return "AnyStr"
-
-    def __repr__(self) -> str:
-        return "AnyStr"
-
-
-class _Anything:
-    def __eq__(self, other: object) -> bool:
-        return True
-
-    def __hash__(self) -> int:
-        return id(ANYTHING)
-
-    def __str__(self) -> str:
-        return "Anything"
-
-    def __repr__(self) -> str:
-        return "Anything"
-
-
-ANY_INT = _AnyInt()
-ANY_STR = _AnyStr()
-ANYTHING = _Anything()
